requirements.txt
pytorch
torchvision
lmdb
